Remove expect classes.
