+++ src/commonMain/kotlin/cz/drekorian/avonmobilefetcher/CsvPrinter.kt
@@ -3,18 +3,10 @@ package cz.drekorian.avonmobilefetcher
 internal const val CSV_SEPARATOR = ";"
 
 /**
- * This class writes Excel-compatible CSV files into a given file path.
+ * Prints given [data] into a file with file path [filePath].
  *
- * @author Marek Osvald
+ * @param filePath file path to store the given [data]
+ * @param data data to store
+ * @param header a header to write at the beginning of the CSV
  */
-expect object CsvPrinter {
-
-    /**
-     * Prints given [data] into a file with file path [filePath].
-     *
-     * @param filePath file path to store the given [data]
-     * @param data data to store
-     * @param header a header to write at the beginning of the CSV
-     */
-    fun print(filePath: String, data: Iterable<String>, vararg header: String)
-}
+expect fun printCsv(filePath: String, data: Iterable<String>, vararg header: String)
