+++ src/commonMain/kotlin/cz/drekorian/avonmobilefetcher/multiplatform/util/NCalendar.kt
@@ -5,6 +5,7 @@ package cz.drekorian.avonmobilefetcher.multiplatform.util
  *
  * @property year calendar year (e.g., 2022)
  * @property month calendar month (e.g., 1 for January, 12 for December)
+ * @author Marek Osvald
  */
 data class NCalendar(
     val year: Int,
@@ -12,16 +13,9 @@ data class NCalendar(
 )
 
 /**
- * This class provides [NCalendar] instances.
+ * Returns a new [NCalendar] instance representing the current date.
  *
+ * @return [NCalendar] instance representing the current date.
  * @author Marek Osvald
  */
-expect object NCalendarProvider {
-
-    /**
-     * Returns a new [NCalendar] instance representing the current date.
-     *
-     * @return [NCalendar] instance representing the current date.
-     */
-    fun getInstance(): NCalendar
-}
+expect fun getCalendarInstance(): NCalendar
