+++ src/jvmMain/kotlin/cz/drekorian/avonmobilefetcher/multiplatform/util/NCalendar.kt
@@ -2,15 +2,12 @@ package cz.drekorian.avonmobilefetcher.multiplatform.util
 
 import java.util.Calendar
 
-actual object NCalendarProvider {
-
-    actual fun getInstance(): NCalendar {
-        val calendar = Calendar.getInstance()
-        val year = calendar[Calendar.YEAR]
-        val month = calendar[Calendar.MONTH] + 1
-        return NCalendar(
-            year = year,
-            month = month,
-        )
-    }
+actual fun getCalendarInstance(): NCalendar {
+    val calendar = Calendar.getInstance()
+    val year = calendar[Calendar.YEAR]
+    val month = calendar[Calendar.MONTH] + 1
+    return NCalendar(
+        year = year,
+        month = month,
+    )
 }
