+++ src/jvmMain/kotlin/cz/drekorian/avonmobilefetcher/multiplatform/util/NLocaleProvider.kt
@@ -2,14 +2,11 @@ package cz.drekorian.avonmobilefetcher.multiplatform.util
 
 import java.util.Locale
 
-actual object NLocaleProvider {
+actual val UK: NLocale = Locale.UK.toNLocale()
 
-    actual val UK: NLocale = Locale.UK.toNLocale()
+actual fun getDefaultLocale(): NLocale = Locale.getDefault().toNLocale()
 
-    actual fun getDefault(): NLocale = Locale.getDefault().toNLocale()
-
-    private fun Locale.toNLocale(): NLocale = NLocale(
-        languageCode = language,
-        region = country,
-    )
-}
+private fun Locale.toNLocale(): NLocale = NLocale(
+    languageCode = language,
+    region = country,
+)
