+++ src/nativeMain/kotlin/cz/drekorian/avonmobilefetcher/multiplatform/util/NCalendar.kt
@@ -1,0 +1,16 @@
+package cz.drekorian.avonmobilefetcher.multiplatform.util
+
+internal val months = mapOf(
+    "Jan" to 1,
+    "Feb" to 2,
+    "Mar" to 3,
+    "Apr" to 4,
+    "May" to 5,
+    "Jun" to 6,
+    "Jul" to 7,
+    "Aug" to 8,
+    "Sep" to 9,
+    "Oct" to 10,
+    "Nov" to 11,
+    "Dec" to 12,
+)
