+++ src/nativeMain/kotlin/cz/drekorian/avonmobilefetcher/multiplatform/util/NLocaleProvider.kt
@@ -2,16 +2,13 @@ package cz.drekorian.avonmobilefetcher.multiplatform.util
 
 import platform.windows.GetUserDefaultLangID
 
-actual object NLocaleProvider {
+private const val LANG_ID_CZECH = 1029
+private const val LANG_ID_UK = 2057
 
-    private const val LANG_ID_CZECH = 1029
-    private const val LANG_ID_UK = 2057
+actual val UK: NLocale = NLocale("en", "GB")
 
-    actual val UK: NLocale = NLocale("en", "GB")
-
-    actual fun getDefault(): NLocale = when (GetUserDefaultLangID().toInt()) {
-        LANG_ID_CZECH -> NLocale("cs", "CZ")
-        LANG_ID_UK -> UK
-        else -> UK
-    }
+actual fun getDefaultLocale(): NLocale = when (GetUserDefaultLangID().toInt()) {
+    LANG_ID_CZECH -> NLocale("cs", "CZ")
+    LANG_ID_UK -> UK
+    else -> UK
 }
